backend/requirements.txt
django>=3.1
requests>=2.23.0
django-cors-middleware>=1.3.1
django-debug-toolbar==2.0
django-cors-headers>=3.2.1
djangorestframework>=3.11.0
djangorestframework-simplejwt>=4.3.0
sentry-sdk==0.14.4
celery>=4.4.5
django-redis>=4.12.1
redis>=3.5.3
future>=0.18.2
eventlet>=0.25.2
pyotp>=2.3.0
gunicorn>=19.9.0
Pillow>=6.2.1
mysqlclient>=2.0.1
python-dateutil>=2.8.1